Move model methods into their own files.
