+++ nimare/meta/cbma/__init__.py
@@ -4,5 +4,8 @@ Top-level namespace for meta-analyses.
 
 from .ale import ALE, SCALE
 from .mkda import MKDAChi2, MKDADensity, KDA
-from .model import BHICP, HPGRF, SBLFR, SBR
+from .sblfr import SBLFR
+from .bhicp import BHICP
+from .hpgrf import HPGRF
+from .sbr import SBR
 from .kernel import ALEKernel, MKDAKernel, KDAKernel
