+++ nimare/meta/cbma/bhicp.py
@@ -1,0 +1,18 @@
+"""
+Bayesian hierarchical cluster process model
+"""
+from .base import CBMAEstimator
+from ...due import due, Doi
+
+
+@due.dcite(Doi('10.1198/jasa.2011.ap09735'),
+           description='Introduces the BHICP model.')
+class BHICP(CBMAEstimator):
+    """
+    Bayesian hierarchical cluster process model
+    """
+    def __init__(self, dataset):
+        self.dataset = dataset
+
+    def fit(self, sample):
+        pass
